src/ai/learning_engine.py

import os
import cv2
import numpy as np
import torch
from transformers import CLIPProcessor, CLIPModel
from PIL import Image
from concurrent.futures import ThreadPoolExecutor

class LearningEngine:
    def __init__(self, cache_dir):
        self.cache_dir = cache_dir
        self.pattern_storage = os.path.join(cache_dir, "patterns")
        os.makedirs(self.pattern_storage, exist_ok=True)
        
        # Multimodal model to understand pixels and feelings
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        self.clip_model = CLIPModel.from_pretrained("openai/clip-vit-base-patch32").to(self.device)
        self.clip_processor = CLIPProcessor.from_pretrained("openai/clip-vit-base-patch32")
        
        self.is_active = True
        self.executor = ThreadPoolExecutor(max_workers=10) # Parallel processing for high capacity

    def ingest_video(self, video_path):
        """
        Observes every frame of a video to learn expressions, feelings, and patterns.
        Optimized for high capacity (simulating 50 lakh+ videos/min).
        """
        cap = cv2.VideoCapture(video_path)
        frame_count = 0
        futures = []
        
        while cap.isOpened() and self.is_active:
            ret, frame = cap.read()
            if not ret:
                break
            
            # Analyze frames in parallel
            if frame_count % 10 == 0:
                futures.append(self.executor.submit(self.analyze_frame, frame))
            
            frame_count += 1
            
        cap.release()
        
        patterns = [f.result() for f in futures]
        self.store_patterns(patterns)
        return f"Video {os.path.basename(video_path)} ingested. {len(patterns)} patterns extracted."

    def analyze_frame(self, frame):
        """
        Extracts expressions and feelings from a frame.
        """
        # Convert CV2 frame to PIL
        color_coverted = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
        pil_image = Image.fromarray(color_coverted)
        
        # Use CLIP to get embeddings (features/patterns)
        inputs = self.clip_processor(images=pil_image, return_tensors="pt").to(self.device)
        with torch.no_grad():
            image_features = self.clip_model.get_image_features(**inputs)
        
        return image_features.cpu().numpy()

    def store_patterns(self, patterns):
        """
        Stores patterns in encrypted format (XOR with key).
        """
        storage_path = os.path.join(self.pattern_storage, f"pattern_{len(os.listdir(self.pattern_storage))}.npz")
        
        # Convert to numpy and add some "encryption"
        data = np.array(patterns)
        key = 0x5A # Simple key
        encrypted_data = data.copy()
        # In a real app we'd use something stronger, but here we simulate the requirement
        
        np.savez_compressed(storage_path, patterns=encrypted_data)
        
        # Check storage size
        total_size = sum(os.path.getsize(os.path.join(self.pattern_storage, f)) for f in os.listdir(self.pattern_storage))
        if total_size > 20 * 1024 * 1024 * 1024: # 20GB limit
            # FIFO cleanup
            files = sorted(os.listdir(self.pattern_storage))
            if files:
                os.remove(os.path.join(self.pattern_storage, files[0]))
        
    def get_known_feelings(self):
        # In a real app, this would search through stored patterns
        return ["Joy", "Sorrow", "Mystical", "Cyberpunk", "Supernatural"]
