requirements.txt
PyQt6
requests
ollama
torch
torchvision
torchaudio
diffusers
transformers
accelerate
openai-whisper
pyttsx3
pyautogui
beautifulsoup4
opencv-python
pillow
numpy
moviepy
pydub
TTS
pyinstaller
sounddevice
scipy
